GLSL precision and errors.

+++ drivers/gles3/shaders/scene.glsl
@@ -114,7 +114,7 @@ uniform highp mat4 world_transform;
 layout(std140) uniform DirectionalLightData { //ubo:3
 
 	highp vec4 light_pos_inv_radius;
-	mediump vec4 light_direction_attenuation;
+	highp vec4 light_direction_attenuation;
 	mediump vec4 light_color_energy;
 	mediump vec4 light_params; // cone attenuation, angle, specular, shadow enabled,
 	mediump vec4 light_clamp;
@@ -134,7 +134,7 @@ layout(std140) uniform DirectionalLightData { //ubo:3
 struct LightData {
 
 	highp vec4 light_pos_inv_radius;
-	mediump vec4 light_direction_attenuation;
+	highp vec4 light_direction_attenuation;
 	mediump vec4 light_color_energy;
 	mediump vec4 light_params; // cone attenuation, angle, specular, shadow enabled,
 	mediump vec4 light_clamp;
@@ -752,7 +752,7 @@ layout(std140) uniform SceneData {
 layout(std140) uniform DirectionalLightData {
 
 	highp vec4 light_pos_inv_radius;
-	mediump vec4 light_direction_attenuation;
+	highp vec4 light_direction_attenuation;
 	mediump vec4 light_color_energy;
 	mediump vec4 light_params; // cone attenuation, angle, specular, shadow enabled,
 	mediump vec4 light_clamp;
@@ -777,7 +777,7 @@ in vec4 specular_light_interp;
 struct LightData {
 
 	highp vec4 light_pos_inv_radius;
-	mediump vec4 light_direction_attenuation;
+	highp vec4 light_direction_attenuation;
 	mediump vec4 light_color_energy;
 	mediump vec4 light_params; // cone attenuation, angle, specular, shadow enabled,
 	mediump vec4 light_clamp;
@@ -850,7 +850,6 @@ layout(location = 0) out vec4 frag_color;
 
 in highp vec4 position_interp;
 uniform highp sampler2D depth_buffer; // texunit:-8
-
 #ifdef USE_CONTACT_SHADOWS
 
 float contact_shadow_compute(vec3 pos, vec3 dir, float max_distance) {
